finger_tracing_refactor/requirements.txt
opencv-python
numpy
pyrealsense2
mediapipe>=0.10.14
PyYAML
scipy
